COM-20 Migrated data table away from @/components/tables/tests
